+++ payload/install-openhtpc-fedora.sh
@@ -2,7 +2,7 @@
 
 set -Eeuo pipefail
 
-readonly OPENHTPC_VERSION="1.1.0-dev37"
+readonly OPENHTPC_VERSION="1.1.0-dev38"
 
 
 readonly SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"